Fixes #71 on two fronts:
The first fix, which was enough by itself, was to remove the hack where
CompImageHDU._prewriteto did not call BinTableHDU._prewriteto.  That
hack was no longer necessary, and removing it allowed the necesary calls
to FITS_rec._scale_back to ensure that the VLA pointers are written
properly.  This also allowed removing some code that was redundant with
BinTableHDU._prewriteto.

The second fix, which was entirely optional, was the introduction of an
internal flag _manages_own_heap.  CompImageHDU has _manages_own_heap =
True.  What this means is basically "I'm going to arrange my heap my
own way, so when writing out don't reorder how data is arranged in the
heap according to the defaults."  This is helpful because when calling
out to CFITSIO it writes the heap data in a particular order, and
there's no reason to waste time reordering that data ourselves when the
order it's already in is fine.

git-svn-id: http://svn6.assembla.com/svn/pyfits/trunk@2764 ed100bfc-0583-0410-97f2-c26b58777a21