Adds two regression tests for #3865, one for the surface issue that a BLANK keyword is invalid when used with float data, and that it should be correctly ignored and the user warned about its invalid use.
The second test is for the deeper issue that BLANK keywords weren't
being removed when auto-rescaling data on read.  It also didn't
properly restore the correct BLANK when using the scale_back=True
option.

Both of these issues are fixed so that the tests pass.

This highlighted other areas where the BLANK keyword is poorly handled
at the user API level, however, which should be addressed in a future
release (namely, that users can't specify how to handle NaNs when
storing floats as ints unless they manually set the correct BLANK).